chore: cleanup cicd
chore: update

chore: fix ut

chore: add telemetry back and fix ut

chore: update
