+++ packages/fx-core/package.json
@@ -9,7 +9,7 @@
     "lint": "eslint \"src/**/*.ts\" \"tests/**/*.ts\"",
     "test:bot": "nyc mocha \"tests/plugins/resource/bot/**/*.test.ts\"",
     "test:botv2": "nyc mocha \"tests/plugins/resource/bot/unit/v2/**/*.test.ts\"",
-    "test:cicd": "nyc mocha \"tests/plugins/resource/cicd/**/*.test.ts\"",
+    "test:cicd": "nyc mocha \"tests/component/feature/cicd/**/*.test.ts\"",
     "test:dotnet": "nyc mocha \"tests/plugins/resource/dotnet/**/*.test.ts\"",
     "test:function": "nyc mocha \"tests/plugins/resource/function/**/*.test.ts\"",
     "test:localdebug": "nyc mocha \"tests/plugins/resource/localdebug/**/*.test.ts\"",
