+++ packages/fx-core/src/component/feature/cicd/errors.ts
@@ -1,7 +1,6 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-import { getLocalizedString } from "../../../common/localizeUtils";
 import { ErrorNames, Suggestions } from "./constants";
 
 export enum ErrorType {
