+++ packages/fx-core/src/component/feature/cicd/provider/enums.ts
