+++ packages/fx-core/src/component/feature/cicd/result.ts
