+++ packages/fx-core/src/component/questionV3.ts
@@ -87,7 +87,6 @@ import { getPlatformRuntime } from "../plugins/resource/bot/v2/mapping";
 import { buildQuestionNode } from "./resource/azureSql/questions";
 import { functionNameQuestion } from "../plugins/resource/function/question";
 import { ApiConnectorImpl } from "./feature/apiconnector/ApiConnectorImpl";
-import { addCicdQuestion } from "./feature/cicd";
 import { BuiltInFeaturePluginNames } from "../plugins/solution/fx-solution/v3/constants";
 import {
   frameworkQuestion,
@@ -99,6 +98,7 @@ import { Constants } from "../plugins/resource/aad/constants";
 import { getQuestionsForDeployAPIM } from "./resource/apim";
 import { canAddSso } from "./feature/sso";
 import { getAddSPFxQuestionNode } from "./feature/spfx";
+import { addCicdQuestion } from "./feature/cicd/cicd";
 
 export async function getQuestionsForProvisionV3(
   context: v2.Context,
