+++ packages/fx-core/src/plugins/resource/index.ts
@@ -14,4 +14,3 @@ export * from "./simpleauth";
 export * from "./spfx";
 export * from "./sql";
 export * from "./keyvault";
-export * from "./cicd";
