+++ packages/fx-core/src/plugins/solution/fx-solution/ResourcePluginContainer.ts
@@ -39,7 +39,6 @@ export const ResourcePluginsV2 = {
   AppStudioPlugin: "AppStudioPluginV2",
   SimpleAuthPlugin: "SimpleAuthPluginV2",
   KeyVaultPlugin: "KeyVaultPluginV2",
-  CICDPlugin: "CICDPluginV2",
 };
 
 /**
